only pre-fill related objects if some objects were received
This fixes, in a roundabout manner, a bug with Ambari 1.7 returning
and empty alert_definitions on the cluster objects, when alert_definitions
were added in 2.0 as related objects.  Since it tried to populate the related
collection, it failed the version check and threw an exception on every
cluster GET call.